services/api/requirements.txt
fastapi
uvicorn
pillow
python-multipart
pydantic
httpx
numpy
scipy
opencv-python-headless